Databse work: added add recipe, search recipe, and return JSON string methods in the index.js file for the database. Implemented new testing as well, in test.js.
Database work: added add recipe, search recipe, and return JSON string methods in the index.js file for the database.

Implemented new testing as well, in test.js.  Made sure that the tests are passing. I also manually verified these by using curl to send a POST request to the endpoint, and made sure the output in the terminal matched expectations.

I still need to fix the port error, which is flagging in the test,js file, but is not directly related to the new methods implemented. It is an "uncaught error outside testsuite."

This is on a new branch called database-work-v2